Use kubernetes network for management
If no controllers are present, currently fuel may allocate the .1 on the
management network which can lead to weird results.  So for now if no
controllers are in the environment, then use the kubernetes network for
management functions.

Also update flannel options to match the other classes and remove the
extra - before -etcd-servers